Soft destroy dependents
The dependent destroy callback mechanism in Rails calls destroy! instead
of destroy for has_many associations. With the existing implementation
of acts_as_paranoid, this would unintentionally cause dependent objects
to be hard-destroyed. This change makes destroy and destroy! identical,
and adds a separate destroy_fully! method for full destroy.