Make paranoid_column_reference a method
* paranoid_column_reference is currently a class attribute which is set when a class is
loaded and the acts_as_paranoid method is invoked. A
rails model can set its table_name after it declares itself as
acts_as_paranoid.  This means that whenever paranoid_column_reference is
referenced in the gem it does not match the table_name for the model
which can cause errors

e.g.

```ruby
class User
  self.table_name = "Account"
  acts_as_paranoid
end

class User
  acts_as_paranoid
  self.table_name = "Account"
end
```

This change turns this attribute into a method so that it will always
match whatever the table_name is.