Allow model to set table_name after acts_as_paranoid macro
* paranoid_column_reference is currently a class attribute which is set
when a class is loaded and the acts_as_paranoid method is invoked. A
rails model can set its table_name after it declares itself as
acts_as_paranoid. This means that whenever paranoid_column_reference is
referenced in the gem it does not match the table_name for the model
which can cause errors.

* e.g. ActsAsParanoid::Core#only_deleted references
`paranoid_column_reference` so in the below examples when we call this
method on the paranoid class it tries to look up a table_name matching
the model name. If the table doesn't exist a sql error is raised because
the table does not exist. If the table does exist (maybe you're moving
data to a new table and deprecating old ones) you'll end up returning
the old records.

* This change makes it a method so that it returns whatever the current
value of table_name is

```ruby

class User
  acts_as_paranoid
  self.table_name = "Account"
end

User.paranoid_column_reference
=> "user.deleted_at"
```

```ruby

class User
  acts_as_paranoid
  self.table_name = "Account"
end

User.paranoid_column_reference
=> "account.deleted_at"
```